requirement.txt
beautifulsoup4
pycryptodome
schedule
streamlit